GDCM 2017-09-15 (87e3c363)
Code extracted from:

    http://git.code.sf.net/p/gdcm/gdcm.git

at commit 87e3c363004ef68545f6d724c8f0ddb251cc44b4 (87e3c363).

+++ Source/DataStructureAndEncodingDefinition/gdcmVM.cxx
@@ -348,9 +348,9 @@ VM::VMType VM::GetVMTypeFromLength(unsigned int length, unsigned int size)
     }
 }
 
-unsigned int VM::GetNumberOfElementsFromArray(const char *array, unsigned int length)
+size_t VM::GetNumberOfElementsFromArray(const char *array, size_t length)
 {
-  unsigned int c=0;
+  size_t c=0;
   if( !length || !array ) return 0;
   const char *parray = array;
   const char *end = array + length;
